requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cffi==1.14.6
cryptography==3.4.7
decouple==0.0.7
Django==3.2.5
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-environ==0.4.5
django-js-asset==1.2.2
django-rest-knox==4.1.0
django-shell-plus==1.1.5
djangorestframework==3.12.4
geographiclib==1.52
geopy==2.2.0
mysqlclient==2.1.0
Pillow==8.3.1
pycparser==2.21
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
